refactor(GeneralInformationForm): update path impotation
